integrate callserver example
