--- examples/04_callserver/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "waku-example",
-  "version": "0.1.0",
-  "type": "module",
-  "private": true,
-  "scripts": {
-    "dev": "waku dev --with-ssr",
-    "build": "waku build --with-ssr",
-    "start": "waku start --with-ssr"
-  },
-  "dependencies": {
-    "react": "18.3.0-canary-b30030471-20240117",
-    "react-dom": "18.3.0-canary-b30030471-20240117",
-    "react-server-dom-webpack": "18.3.0-canary-b30030471-20240117",
-    "waku": "0.19.0"
-  },
-  "devDependencies": {
-    "@types/react": "18.2.48",
-    "@types/react-dom": "18.2.18",
-    "typescript": "5.3.3"
-  }
-}
