--- examples/04_callserver/src/components/funcs.ts
@@ -1,3 +1,0 @@
-'use server';
-
-export const greet = (name: string) => `Hello ${name} from server!`;
