--- examples/04_callserver/src/entries.tsx
@@ -1,31 +1,0 @@
-import { lazy } from 'react';
-import { defineEntries } from 'waku/server';
-import { Slot } from 'waku/client';
-
-const App = lazy(() => import('./components/App.js'));
-
-export default defineEntries(
-  // renderEntries
-  async (input) => {
-    return {
-      App: <App name={input || 'Waku'} />,
-    };
-  },
-  // getBuildConfig
-  async () => [{ pathname: '/', entries: [{ input: '', isStatic: true }] }],
-  // getSsrConfig
-  async (pathname, { isPrd }) => {
-    if (isPrd) {
-      return null;
-    }
-    switch (pathname) {
-      case '/':
-        return {
-          input: '',
-          body: <Slot id="App" />,
-        };
-      default:
-        return null;
-    }
-  },
-);
