--- examples/04_callserver/src/main.tsx
@@ -1,17 +1,0 @@
-import { StrictMode } from 'react';
-import { createRoot, hydrateRoot } from 'react-dom/client';
-import { Root, Slot } from 'waku/client';
-
-const rootElement = (
-  <StrictMode>
-    <Root>
-      <Slot id="App" />
-    </Root>
-  </StrictMode>
-);
-
-if (import.meta.env.WAKU_HYDRATE) {
-  hydrateRoot(document.body, rootElement);
-} else {
-  createRoot(document.body).render(rootElement);
-}
