--- examples/04_callserver/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true,
-    "target": "esnext",
-    "downlevelIteration": true,
-    "esModuleInterop": true,
-    "module": "nodenext",
-    "skipLibCheck": true,
-    "noUncheckedIndexedAccess": true,
-    "exactOptionalPropertyTypes": true,
-    "types": ["react/experimental"],
-    "jsx": "react-jsx",
-  },
-}
