+++ package.json
@@ -15,7 +15,6 @@
     "examples:dev:01_template": "NAME=01_template pnpm run examples:dev",
     "examples:dev:03_minimal": "NAME=03_minimal pnpm run examples:dev",
     "examples:dev:04_promise": "NAME=04_promise pnpm run examples:dev",
-    "examples:dev:04_callserver": "NAME=04_callserver pnpm run examples:dev",
     "examples:dev:05_actions": "NAME=05_actions pnpm run examples:dev",
     "examples:dev:06_nesting": "NAME=06_nesting pnpm run examples:dev",
     "examples:dev:07_router": "NAME=07_router pnpm run examples:dev",
@@ -28,10 +27,9 @@
     "examples:build": "(cd ./examples/${NAME} && pnpm run build)",
     "examples:prd": "pnpm run examples:build && (cd ./examples/${NAME} && pnpm start)",
     "examples:prd:01_template": "NAME=01_template pnpm run examples:prd",
-    "examples:prd:02_minimal": "NAME=02_minimal pnpm run examples:prd",
-    "examples:prd:03_promise": "NAME=03_promise pnpm run examples:prd",
-    "examples:prd:04_callserver": "NAME=04_callserver pnpm run examples:prd",
-    "examples:prd:05_mutation": "NAME=05_mutation pnpm run examples:prd",
+    "examples:prd:03_minimal": "NAME=03_minimal pnpm run examples:prd",
+    "examples:prd:04_promise": "NAME=04_promise pnpm run examples:prd",
+    "examples:prd:05_actions": "NAME=05_actions pnpm run examples:prd",
     "examples:prd:06_nesting": "NAME=06_nesting pnpm run examples:prd",
     "examples:prd:07_router": "NAME=07_router pnpm run examples:prd",
     "examples:prd:08_cookies": "NAME=08_cookies pnpm run examples:prd",
